+++ staticd/.gitignore
@@ -1,0 +1,2 @@
+libstatic.a
+staticd
