+++ staticd/Makefile
@@ -1,0 +1,10 @@
+all: ALWAYS
+	@$(MAKE) -s -C .. staticd/staticd
+%: ALWAYS
+	@$(MAKE) -s -C .. staticd/$@
+
+Makefile:
+	#nothing
+ALWAYS:
+.PHONY: ALWAYS makefiles
+.SUFFIXES:
