+++ staticd/static_vrf.h
@@ -1,0 +1,38 @@
+/*
+ * STATICd - vrf header
+ * Copyright (C) 2018 Cumulus Networks, Inc.
+ *               Donald Sharp
+ *
+ * FRR is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2, or (at your option) any
+ * later version.
+ *
+ * FRR is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; see the file COPYING; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
+ */
+#ifndef __STATIC_VRF_H__
+#define __STATIC_VRF_H__
+
+struct static_vrf {
+	struct vrf *vrf;
+
+	struct route_table *stable[AFI_MAX][SAFI_MAX];
+};
+
+struct static_vrf *static_vrf_lookup_by_name(const char *vrf_name);
+struct static_vrf *static_vrf_lookup_by_id(vrf_id_t vrf_id);
+
+int static_vrf_has_config(struct static_vrf *svrf);
+
+void static_vrf_init(void);
+
+struct route_table *static_vrf_static_table(afi_t afi, safi_t safi,
+					    struct static_vrf *svrf);
+#endif
