+++ staticd/staticd.conf.sample
@@ -1,0 +1,3 @@
+!
+!
+log stdout
