+++ vtysh/Makefile.am
@@ -146,6 +146,10 @@ vtysh_scan += $(top_srcdir)/pbrd/pbr_vty.c
 vtysh_scan += $(top_srcdir)/pbrd/pbr_debug.c
 endif
 
+if STATICD
+vtysh_scan += $(top_srcdir)/staticd/static_vty.c
+endif
+
 vtysh_cmd_FILES = $(vtysh_scan) \
 		  $(top_srcdir)/lib/keychain.c $(top_srcdir)/lib/routemap.c \
 		  $(top_srcdir)/lib/filter.c $(top_srcdir)/lib/plist.c \
